requirements.txt
openai>=1.30.0
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
tavily-python>=0.4.0
python-dotenv>=1.0.0
streamlit>=1.35.0
llama-index>=0.10.0
llama-index-embeddings-openai>=0.1.0
sentence-transformers>=2.6.0
llama-index-retrievers-bm25>=0.1.0
